Initial Refactoring - Updated folder structure
